requirements.txt
Django==5.0.6
djangorestframework==3.15.1
psycopg2==2.9.9
celery[redis]==5.4.0
django-redis==5.4.0
django-celery-results==2.5.1
flower==2.0.1
celery-singleton==0.3.1


